requirements.txt
aiofiles==0.7.0
aiogram==2.14.3
aiohttp==3.7.4.post0
asgiref==3.4.1
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.6
Django==3.2.7
gunicorn==20.1.0
idna==3.2
iniconfig==1.1.1
loguru==0.5.3
multidict==5.1.0
packaging==21.0
pluggy==1.0.0
psycopg2==2.9.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.5
pytest-asyncio==0.15.1
pytest-django==4.4.0
python-dotenv==0.19.0
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
toml==0.10.2
typing-extensions==3.10.0.2
urllib3==1.26.7
yarl==1.6.3
